Refactor ShopPage.tsx: Extract components and hooks for better code organization
- Reduced ShopPage.tsx from 1309 lines to 751 lines (42% reduction)
- Extracted 6 reusable components:
  - FilterSidebar: Desktop filter panel with search
  - ResultsHeader: Results count, view toggle, active filters
  - CardDisplayArea: Card grid/list display with grouping logic
  - MobileFilterModal: Mobile filter overlay with accessibility
  - MobileFilterButton: Mobile filter toggle
  - ShopHeader: Enhanced header with cart and currency selector
- Created 3 new hooks:
  - useCardFetching: Complete API integration for card data
  - useShopViewMode: View preferences and currency management
  - useShopKeyboardShortcuts: Keyboard shortcut handling
- Enhanced existing components:
  - ShopHeader: From placeholder to full implementation
  - useCardFetching: From stub to complete implementation
- Updated barrel exports for clean imports
- Maintains 100% backward compatibility
- Improves code reusability, testability, and maintainability
