+++ apps/web/src/features/shop/index.ts
@@ -1,5 +1,20 @@
 export { default as ShopPage } from './ShopPage';
+
+// Hooks
 export * from './hooks/useCart';
 export * from './hooks/useShopFilters';
 export * from './hooks/useRecentlyViewed';
+export * from './hooks/useVariationSelection';
+export * from './hooks/useCardFetching';
+export * from './hooks/useShopViewMode';
+export * from './hooks/useShopKeyboardShortcuts';
+
+// Components
+export { default as ShopHeader } from './components/ShopHeader';
+export { default as FilterSidebar } from './components/FilterSidebar';
+export { default as ResultsHeader } from './components/ResultsHeader';
+export { default as CardDisplayArea } from './components/CardDisplayArea';
+export { default as MobileFilterModal } from './components/MobileFilterModal';
+export { default as MobileFilterButton } from './components/MobileFilterButton';
+
 export * from './components/Cart'
